update ui and files
major changes in files and format, working on the profile picture
