+++ app/src/main/AndroidManifest.xml
@@ -14,8 +14,6 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.MemeIt">
-        <activity android:name=".Auth.EditProfile"></activity>
-        <activity android:name=".Profile" />
         <activity
             android:name=".Auth.Register"
             android:label="@string/title_activity_login"
