+++ app/src/main/java/com/example/memeit/fragments/ProfileFragment.java
@@ -1,6 +1,9 @@
 package com.example.memeit.fragments;
 
 import android.content.Intent;
+import android.graphics.Bitmap;
+import android.graphics.BitmapFactory;
+import android.net.Uri;
 import android.os.Bundle;
 import android.util.Log;
 import android.view.LayoutInflater;
@@ -13,19 +16,32 @@ import android.widget.Toast;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
+import androidx.appcompat.app.AlertDialog;
 import androidx.fragment.app.Fragment;
 
-import com.example.memeit.Auth.EditProfile;
 import com.example.memeit.Auth.Login;
 import com.example.memeit.R;
-import com.google.android.material.bottomnavigation.BottomNavigationView;
+import com.parse.FindCallback;
+import com.parse.GetCallback;
 import com.parse.LogOutCallback;
 import com.parse.ParseException;
+import com.parse.ParseFile;
+import com.parse.ParseObject;
+import com.parse.ParseQuery;
 import com.parse.ParseUser;
 
+import java.util.List;
+
 public class ProfileFragment extends Fragment {
-    Button changeemail, changename, changepass, logout;
-    EditText setName, setEmail, setPass, curPass;
+    public static final String TAG = "ProfileFragment";
+    private AlertDialog.Builder dialogbuilder;
+    private AlertDialog dialog;
+    Button changeemail, changepass, savenewpassword, saveemail, changeuser, saveuser, logout;
+    private EditText newpassword, confirmnewpassword;
+    private EditText newemail, confirmnewemail;
+    private EditText newuser, confirmuser;
+    private ImageView profileimage;
+    EditText setName, setEmail;
     ParseUser parseUser = ParseUser.getCurrentUser();
 
     public ProfileFragment(){
@@ -46,12 +62,11 @@ public class ProfileFragment extends Fragment {
         ParseUser currentUser = ParseUser.getCurrentUser();
         setEmail = view.findViewById(R.id.setEmail);
         setName = view.findViewById(R.id.setName);
-        setPass = view.findViewById(R.id.setPass);
-        curPass = view.findViewById(R.id.curPass);
 
         changeemail= view.findViewById(R.id.changeemail);
-        changename= view.findViewById(R.id.changename);
+        changeuser= view.findViewById(R.id.changename);
         changepass= view.findViewById(R.id.changepass);
+        profileimage = view.findViewById(R.id.profileimage);
 
         if (currentUser != null) {
             // Name, email address, and profile photo Url
@@ -59,6 +74,36 @@ public class ProfileFragment extends Fragment {
             setName.setText(name);
             String email = currentUser.getEmail();
             setEmail.setText(email);
+
+            // profile image
+            ParseQuery<ParseUser> imageQuery = ParseQuery.getQuery("Images");
+            imageQuery.whereEqualTo("objectId",ParseUser.getCurrentUser().getObjectId());
+            imageQuery.findInBackground(new FindCallback<ParseObject>()
+            {
+                @Override
+                public void done(List<ParseUser> users,ParseException e)
+                {
+                    for(ParseUser user : users)
+                    {
+                        ParseFile UserProPicFile = object.getParseFile("ImageColumnName");
+                        byte[] byteArray = new byte[0];
+
+                        byteArray = UserProPicFile.getData();
+                        Bitmap ProselectBit = BitmapFactory.decodeByteArray(byteArray,0,byteArray.length); //Bitmap with [rofile picture
+                    }
+                }
+            });
+//            imageQuery.getFirstInBackground(new GetCallback<ParseObject>() {
+//                public void done(ParseObject Image, ParseException e) {
+//                    if (e == null) {
+//                        ParseFile postImage = object.getParseFile(ParseConstants.PARSE_KEY_FILE);
+//                        String imageUrl = postImage.getUrl() ;//live url
+//                        Uri imageUri = Uri.parse(imageUrl);
+//                    } else {
+//                        // Something is wrong
+//                    }
+//                }
+//            });
         }
 
         logout= view.findViewById(R.id.logout);
@@ -72,51 +117,157 @@ public class ProfileFragment extends Fragment {
                             return;
                         }
                         Intent i= new Intent(getActivity(), Login.class);
-                        Toast.makeText(getContext(), "Logged out!", Toast.LENGTH_SHORT).show();
+                        Toast.makeText(getContext(), "Logged out", Toast.LENGTH_SHORT).show();
                         startActivity(i);
                     }
                 });
             }
         });
 
+        changeuser.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                createNewContactDialoguser();
+            }
+        });
+
         changepass.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                String newpassword= setPass.getText().toString();
-                // check the length of password
-                if(newpassword.length()<6){
-                    Toast.makeText(getContext(),"Password Must be >= 6 Characters",Toast.LENGTH_SHORT).show();
-                }
-                if(newpassword.isEmpty()){
-                    Toast.makeText(getContext(),"Password cannot be empty",Toast.LENGTH_SHORT).show();
-                }
-                else {
-                    parseUser.setPassword(newpassword);
-                    Toast.makeText(getContext(), "Password Updated", Toast.LENGTH_SHORT).show();
-                    setPass.setText("");
-                    curPass.setText("");
-                }
+                createNewContactDialogpass();
             }
         });
 
         changeemail.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                String newemail= setEmail.getText().toString();
-                if(newemail.isEmpty()){
-                    Toast.makeText(getContext(), "Email can't be empty!", Toast.LENGTH_SHORT).show();
+                createNewContactDialogemail();
+            }
+        });
+    }
+
+    // user
+    public void createNewContactDialoguser(){
+        dialogbuilder = new AlertDialog.Builder(getContext());
+        final View conactPopupView = getLayoutInflater().inflate(R.layout.popupuser, null);
+        newuser = (EditText) conactPopupView.findViewById(R.id.newuser);
+        confirmuser = (EditText) conactPopupView.findViewById(R.id.confirmuser);
+        saveuser  = (Button) conactPopupView.findViewById(R.id.saveuser);
+        // get the old value of old username
+        String olduser = ParseUser.getCurrentUser().getUsername();
+        dialogbuilder.setView(conactPopupView);
+        dialog = dialogbuilder.create();
+        dialog.show();
+
+        saveuser.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                if(newuser.getText().toString().equals(confirmuser.getText().toString())){
+                    // check if the old user name is being used
+                    if(olduser.equals(newuser.getText().toString())){
+                        newuser.setError("Cannot use old username");
+                    }
+                    // check the length of username
+                    if(newuser.length()==0){
+                        newuser.setError("Username Cannot be empty");
+                    }
+                    else {
+                        parseUser.setUsername(newuser.getText().toString());
+                        parseUser.saveInBackground();
+                        Toast.makeText(getActivity(), "Username Updated", Toast.LENGTH_SHORT).show();
+                        setName.setText(newuser.getText().toString());
+                        dialog.dismiss();
+                    }
                 }
                 else{
-                    parseUser.setEmail(newemail);
-                    parseUser.saveInBackground();
-                    Toast.makeText(getContext(),"Email Updated",Toast.LENGTH_SHORT).show();
-                    setEmail.setText("");
+                    Toast.makeText(getActivity(),"Username failed to Updated" ,Toast.LENGTH_SHORT).show();
+                    dialog.dismiss();
                 }
             }
         });
-
-
-
     }
 
+    // pass
+    public void createNewContactDialogpass(){
+        dialogbuilder = new AlertDialog.Builder(getContext());
+        final View conactPopupView = getLayoutInflater().inflate(R.layout.popuppassword, null);
+        newpassword = (EditText) conactPopupView.findViewById(R.id.newpassword);
+        confirmnewpassword = (EditText) conactPopupView.findViewById(R.id.confirmnewpassword);
+        savenewpassword  = (Button) conactPopupView.findViewById(R.id.savepassword);
+        parseUser = ParseUser.getCurrentUser();
+
+        dialogbuilder.setView(conactPopupView);
+        dialog = dialogbuilder.create();
+        dialog.show();
+
+        savenewpassword.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                if(newpassword.getText().toString().equals(confirmnewpassword.getText().toString())){
+                    // check the length of password
+                    String password =  newpassword.getText().toString().trim();
+                    Log.i("lol", String.valueOf(password.length()));
+                    if(password.length()<6){
+                        newpassword.setError("Password Must be >= 6 Characters");
+                    }
+                    if(newpassword.getText().toString().isEmpty()){
+                        newpassword.setError("Password cannot be empty");
+                    }
+                    else {
+                        parseUser.setPassword(newpassword.getText().toString());
+                        parseUser.saveInBackground();
+                        Toast.makeText(getActivity(), "Password Updated", Toast.LENGTH_SHORT).show();
+                        dialog.dismiss();
+                    }
+                }
+                else{
+                    Toast.makeText(getActivity(),"Password failed to Updated" ,Toast.LENGTH_SHORT).show();
+                    dialog.dismiss();
+                }
+            }
+        });
+    }
+
+
+
+
+    // email
+    public void createNewContactDialogemail(){
+        dialogbuilder = new AlertDialog.Builder(getContext());
+        final View conactPopupView = getLayoutInflater().inflate(R.layout.popupemail, null);
+        newemail = (EditText) conactPopupView.findViewById(R.id.newuser);
+        confirmnewemail = (EditText) conactPopupView.findViewById(R.id.confirmuser);
+        saveemail  = (Button) conactPopupView.findViewById(R.id.saveuser);
+        // get the old value of old email
+        String oldemail = ParseUser.getCurrentUser().getEmail();
+        dialogbuilder.setView(conactPopupView);
+        dialog = dialogbuilder.create();
+        dialog.show();
+
+        saveemail.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                if(newemail.getText().toString().equals(confirmnewemail.getText().toString())){
+                    if(oldemail.equals(newemail.getText().toString())){
+                        newemail.setError("Cannot use old email");
+                    }
+                    if(newemail.getText().toString().isEmpty()){
+                        newemail.setError("Email cannot be empty");
+                    }
+                    else {
+                        parseUser.setEmail(newemail.getText().toString());
+                        parseUser.saveInBackground();
+                        Toast.makeText(getActivity(), "Email Updated", Toast.LENGTH_SHORT).show();
+                        setEmail.setText(newemail.getText().toString());
+                        dialog.dismiss();
+                    }
+                }
+                //shouldn't this part be different, correspond to the password button? This is still in the change email section.
+                else{
+                    Toast.makeText(getActivity(),"Email failed to Updated",Toast.LENGTH_SHORT).show();
+                    dialog.dismiss();
+                }
+            }
+        });
+    }
 }
