--- app/src/main/res/layout/activity_edit_profile.xml
@@ -1,72 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".Auth.EditProfile">
-    <include
-        android:id="@+id/appbar"
-        layout="@layout/toolbar" />
-
-    <Button
-        android:id="@+id/changeemail"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_below="@+id/profileimage"
-        android:layout_marginHorizontal="110dp"
-        android:layout_marginTop="50dp"
-        android:text="change email"
-        tools:layout_editor_absoluteX="156dp"
-        tools:layout_editor_absoluteY="492dp" />
-
-    <com.google.android.material.bottomnavigation.BottomNavigationView
-        android:id="@+id/botton_navigation"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentBottom="true"
-        android:layout_marginStart="-1dp"
-        android:layout_marginBottom="-12dp"
-        app:labelVisibilityMode="labeled"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:menu="@menu/menu_bottom_navigation"
-        tools:layout_editor_absoluteX="144dp" />
-
-    <Button
-        android:id="@+id/changename"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_below="@+id/changeemail"
-        android:layout_marginHorizontal="100dp"
-        android:layout_marginTop="50dp"
-        android:text="change password" />
-
-    <Button
-        android:id="@+id/changeuser"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_below="@+id/changepass"
-        android:layout_marginHorizontal="100dp"
-        android:layout_marginTop="50dp"
-        android:text="change username" />
-
-    <ImageView
-        android:id="@+id/profileimage"
-        android:layout_width="154dp"
-        android:layout_height="124dp"
-        android:layout_below="@+id/appbar"
-        android:layout_marginHorizontal="110dp"
-        android:layout_marginTop="100dp"
-        app:srcCompat="@drawable/profileicon" />
-
-    <TextView
-        android:id="@+id/textView3"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_below="@+id/appbar"
-        android:layout_marginHorizontal="115dp"
-        android:layout_marginTop="50dp"
-        android:text="Change Profile Picture" />
-</RelativeLayout>
