--- app/src/main/res/layout/activity_profile.xml
@@ -1,66 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".Profile">
-    <include
-        android:id="@+id/appbar"
-        layout="@layout/toolbar" />
-
-    <Button
-        android:id="@+id/editprofile"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_below="@+id/setEmail"
-        android:layout_marginHorizontal="135dp"
-        android:layout_marginTop="25dp"
-        app:icon="@drawable/settings"
-        app:layout_constraintTop_toTopOf="parent"
-        tools:layout_editor_absoluteX="50dp" />
-
-    <com.google.android.material.bottomnavigation.BottomNavigationView
-        android:id="@+id/botton_navigation"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentBottom="true"
-        android:layout_marginStart="-1dp"
-        android:layout_marginBottom="-12dp"
-        app:labelVisibilityMode="labeled"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:menu="@menu/menu_bottom_navigation"
-        tools:layout_editor_absoluteX="144dp" />
-
-    <EditText
-        android:id="@+id/setName"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_below="@+id/profimage1"
-        android:layout_marginHorizontal="75dp"
-        android:layout_marginTop="50dp"
-        android:ems="10"
-        android:inputType="textPersonName" />
-
-    <EditText
-        android:id="@+id/setEmail"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_below="@+id/setName"
-        android:layout_marginHorizontal="75dp"
-        android:layout_marginTop="23dp"
-        android:ems="10"
-        android:inputType="textEmailAddress" />
-
-    <ImageView
-        android:id="@+id/profimage1"
-        android:layout_width="236dp"
-        android:layout_height="234dp"
-        android:layout_below="@+id/appbar"
-        android:layout_marginHorizontal="75dp"
-        android:layout_marginTop="20dp"
-        app:srcCompat="@drawable/profileicon" />
-
-</RelativeLayout>
