+++ app/src/main/res/layout/popupuser.xml
@@ -10,32 +10,28 @@
         android:id="@+id/saveuser"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_marginStart="136dp"
         android:layout_marginTop="5dp"
         android:text="change user"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/confirmuser" />
+        app:layout_constraintTop_toBottomOf="@+id/confirmuser"
+        tools:layout_editor_absoluteX="136dp" />
 
     <EditText
         android:id="@+id/newuser"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_marginStart="100dp"
-        android:layout_marginTop="172dp"
         android:ems="10"
         android:hint="New User"
         android:inputType="textEmailAddress"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
+        tools:layout_editor_absoluteX="100dp"
+        tools:layout_editor_absoluteY="172dp" />
 
     <EditText
         android:id="@+id/confirmuser"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_marginStart="104dp"
         android:ems="10"
         android:hint="Confirm User"
         android:inputType="textEmailAddress"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/newuser" />
+        app:layout_constraintTop_toBottomOf="@+id/newuser"
+        tools:layout_editor_absoluteX="104dp" />
 </androidx.constraintlayout.widget.ConstraintLayout>
