--- app/src/main/res/layout/toolbar.xml
@@ -1,24 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.appcompat.widget.Toolbar xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:background="@color/colorPrimaryDark"
-    android:elevation="4dp"
-    android:theme="@style/Theme.AppCompat.DayNight.DarkActionBar"
-    android:layout_gravity="end">
-
-    <RelativeLayout
-        android:layout_width="50dp"
-        android:layout_height="match_parent"
-        android:clickable="true"
-        android:gravity="end"/>
-
-    <ImageView
-        android:id="@+id/appbarsignout"
-        android:layout_width="25dp"
-        android:layout_height="25dp"
-        android:layout_gravity="end"
-        android:layout_marginRight="10dp"
-        android:src="@drawable/signout" />
-
-</androidx.appcompat.widget.Toolbar>
